+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add injectable session URL rewrite function to support proxies
 - Add POSTLogger for meeting sessions
 - Integrate POSTLogger into the demo app
+- Add content share integration test
 
 ### Changed
 - Add observer event for content sharing